requirements.txt
asgiref==3.6.0
Django==4.1.8
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
pymongo==3.12.1
pytz==2023.3
sqlparse==0.2.4
levenshtein==0.20.9
unidecode==1.3.2
